staticfied codegen.mdl, also need to revert back to ::
