+++ compiler/src/edu/clemson/resolve/codegen/AbstractCodeGenerator.java
@@ -1,6 +1,6 @@
 package edu.clemson.resolve.codegen;
 
-import edu.clemson.resolve.codegen.model.OutputModelObject;
+import edu.clemson.resolve.codegen.Model.OutputModelObject;
 import edu.clemson.resolve.compiler.AnnotatedModule;
 import edu.clemson.resolve.compiler.ErrorKind;
 import edu.clemson.resolve.RESOLVECompiler;
