+++ compiler/src/edu/clemson/resolve/codegen/JavaCodeGenerator.java
@@ -1,6 +1,6 @@
 package edu.clemson.resolve.codegen;
 
-import edu.clemson.resolve.codegen.model.OutputModelObject;
+import edu.clemson.resolve.codegen.Model.OutputModelObject;
 import edu.clemson.resolve.compiler.AnnotatedModule;
 import edu.clemson.resolve.RESOLVECompiler;
 import edu.clemson.resolve.compiler.ErrorKind;
@@ -42,7 +42,9 @@ class JavaCodeGenerator extends AbstractCodeGenerator {
     }
 
     @Override
-    protected void write(@NotNull ModuleIdentifier moduleIdentifier, @NotNull ST code, @NotNull String outputFileName) {
+    protected void write(@NotNull ModuleIdentifier moduleIdentifier,
+                         @NotNull ST code,
+                         @NotNull String outputFileName) {
         try {
             Writer w = compiler.getOutputFileWriter(moduleIdentifier,
                     outputFileName, new Function<String, File>() {
