+++ compiler/src/edu/clemson/resolve/codegen/ModelConverter.java
@@ -30,8 +30,7 @@
  */
 package edu.clemson.resolve.codegen;
 
-import edu.clemson.resolve.codegen.model.ModelElement;
-import edu.clemson.resolve.codegen.model.OutputModelObject;
+import edu.clemson.resolve.codegen.Model.OutputModelObject;
 import edu.clemson.resolve.compiler.ErrorKind;
 import edu.clemson.resolve.RESOLVECompiler;
 import org.stringtemplate.v4.ST;
