--- compiler/src/edu/clemson/resolve/codegen/model/ConceptImplModule.java
@@ -1,7 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class ConceptImplModule extends AbstractSpecImplModule {
-    public ConceptImplModule(String name, String concept, ModuleFile file) {
-        super(name, concept, file);
-    }
-}
