--- compiler/src/edu/clemson/resolve/codegen/model/Expr.java
@@ -1,7 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-/**
- * Created by daniel on 3/15/15.
- */
-public abstract class Expr extends OutputModelObject {
-}
