--- compiler/src/edu/clemson/resolve/codegen/model/Module.java
@@ -1,46 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-import edu.clemson.resolve.semantics.symbol.ModuleParameterSymbol;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class Module extends OutputModelObject {
-    public String name;
-    public ModuleFile file;
-    @ModelElement
-    public List<FunctionImpl> funcImpls = new ArrayList<>();
-    @ModelElement
-    public List<MemberClassDef> repClasses = new ArrayList<>();
-    @ModelElement
-    public List<VariableDef> memberVars = new ArrayList<>();
-
-    public Module(String name, ModuleFile file) {
-        this.name = name;
-        this.file = file;//who contains us?
-    }
-
-    /**
-     * Like the name suggests, adds getters and member variabes for the formal parameters to a concept
-     * (or enhancement).
-     */
-    public void addGettersAndMembersForModuleParameterSyms(List<ModuleParameterSymbol> symbols) {
-    }
-
-    /**
-     * For implementations that take an operation as a parameter, this method adds both an RType member variable
-     * pointing to the interface wrapping the 'operation' as well as the interior interfaces wrapping calls to the
-     * operation.
-     *
-     * @param wrappedFunction
-     */
-    public void addOperationParameterModelObjects(FunctionDef wrappedFunction) {
-    }
-
-    /** We gen a dummy class for precis to make java happy. Could be filtered too, but this is easier. */
-    public static class PrecisModule extends Module {
-        public PrecisModule(String name, ModuleFile file) {
-            super(name, file);
-        }
-    }
-}
