--- compiler/src/edu/clemson/resolve/codegen/model/OperationParameterDef.java
@@ -1,12 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class OperationParameterDef extends OutputModelObject {
-    @ModelElement
-    public FunctionDef func;
-    public String name;
-
-    public OperationParameterDef(FunctionDef f) {
-        this.func = f;
-        this.name = f.name;
-    }
-}
