--- compiler/src/edu/clemson/resolve/codegen/model/ParameterDef.java
@@ -1,9 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class ParameterDef extends OutputModelObject {
-    public String name;
-
-    public ParameterDef(String name) {
-        this.name = name;
-    }
-}
