--- compiler/src/edu/clemson/resolve/codegen/model/ReturnStat.java
@@ -1,9 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class ReturnStat extends Stat {
-    public String name;
-
-    public ReturnStat(String name) {
-        this.name = name;
-    }
-}
