--- compiler/src/edu/clemson/resolve/codegen/model/TypeInit.java
@@ -1,13 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class TypeInit extends Expr {
-    @ModelElement
-    public Qualifier q;
-    public String typeName, initialValue;
-
-    public TypeInit(Qualifier q, String typeName, String initialValue) {
-        this.q = q;
-        this.typeName = typeName;
-        this.initialValue = initialValue;
-    }
-}
