--- compiler/src/edu/clemson/resolve/codegen/model/TypeInterfaceDef.java
@@ -1,9 +1,0 @@
-package edu.clemson.resolve.codegen.model;
-
-public class TypeInterfaceDef extends OutputModelObject {
-    public String name;
-
-    public TypeInterfaceDef(String name) {
-        this.name = name;
-    }
-}
